Directory change
Moved test files, updated get_database_data helper method, delete old folders
