+++ air-quality-backend/system_tests/utils/cams_utilities.py
@@ -5,10 +5,11 @@ import xarray
 from decimal import Decimal
 from pandas import read_csv
 from pymongo import MongoClient
+from pymongo.errors import ConnectionFailure, OperationFailure
 
 
 def get_dataset_from_coordinates_many_steps(
-    dataset: xarray.Dataset, step: str, latitude: float, longitude: float
+        dataset: xarray.Dataset, step: str, latitude: float, longitude: float
 ) -> xarray.Dataset:
     return dataset.sel(
         indexers={
@@ -21,7 +22,7 @@ def get_dataset_from_coordinates_many_steps(
 
 
 def get_dataset_from_coordinates_single_step(
-    dataset: xarray.Dataset, latitude: float, longitude: float
+        dataset: xarray.Dataset, latitude: float, longitude: float
 ) -> xarray.Dataset:
     return dataset.sel(
         indexers={
@@ -33,12 +34,12 @@ def get_dataset_from_coordinates_single_step(
 
 
 def convert_kg_to_ug(pollutant_value_kg: float) -> float:
-    return float(pollutant_value_kg) * float(10**9)
+    return float(pollutant_value_kg) * float(10 ** 9)
 
 
 def print_cams_data(
-    single_level_dataset_from_coordinates: xarray.Dataset,
-    multi_level_dataset_from_coordinates: xarray.Dataset,
+        single_level_dataset_from_coordinates: xarray.Dataset,
+        multi_level_dataset_from_coordinates: xarray.Dataset,
 ):
     pm2p5 = single_level_dataset_from_coordinates["pm2p5"].values
     pm10 = single_level_dataset_from_coordinates["pm10"].values
@@ -59,11 +60,11 @@ def print_cams_data(
 
 
 def get_raw_cams_data(
-    steps: list[str],
-    single_level_dataset: xarray.Dataset,
-    lat: float,
-    lon: float,
-    multi_level_dataset: xarray.Dataset,
+        steps: list[str],
+        single_level_dataset: xarray.Dataset,
+        lat: float,
+        lon: float,
+        multi_level_dataset: xarray.Dataset,
 ):
     print(
         "\nFetching forecast data from CAMS for latitude: {}, longitude: {}...".format(
@@ -102,30 +103,50 @@ def get_raw_cams_data(
         )
 
 
-def get_database_data(collection_name: str):
+def get_database_data(query: dict, collection_name: str):
     uri = os.environ.get("MONGO_DB_URI")
     db_name = os.environ.get("MONGO_DB_NAME")
-    client = MongoClient(uri)
-    collection = client[db_name][collection_name]
-    cursor = collection.find()
-    database_dictionary_list = []
 
-    for document in cursor:
-        document_as_dict = {
-            "name": document["name"],
-            "created_time": document["created_time"],
-            "last_modified_time": document["last_modified_time"],
-            "forecast_base_time": document["forecast_base_time"],
-            "forecast_range": document["forecast_range"],
-            "forecast_valid_time": document["forecast_valid_time"],
-            "o3_value": document["o3"]["value"],
-            "so2_value": document["so2"]["value"],
-            "no2_value": document["no2"]["value"],
-            "pm10_value": document["pm10"]["value"],
-            "pm2_5_value": document["pm2_5"]["value"],
-        }
-        database_dictionary_list.append(document_as_dict)
-    client.close()
+    if not uri or not db_name:
+        raise ValueError("MONGO_DB_URI and MONGO_DB_NAME must be set!")
+
+    try:
+        client = MongoClient(uri)
+        collection = client[db_name][collection_name]
+        cursor = collection.find(query)
+        database_dictionary_list = []
+
+        for document in cursor:
+            document_as_dict = {
+                "name": document["name"],
+                "created_time": document["created_time"],
+                "last_modified_time": document["last_modified_time"],
+                "forecast_base_time": document["forecast_base_time"],
+                "forecast_range": document["forecast_range"],
+                "forecast_valid_time": document["forecast_valid_time"],
+                "overall_aqi_level": document["overall_aqi_level"],
+                "location_type": document["location_type"],
+                "o3_value": document["o3"]["value"],
+                "so2_value": document["so2"]["value"],
+                "no2_value": document["no2"]["value"],
+                "pm10_value": document["pm10"]["value"],
+                "pm2_5_value": document["pm2_5"]["value"],
+                "o3_aqi_level": document["o3"]["aqi_level"],
+                "so2_aqi_level": document["so2"]["aqi_level"],
+                "no2_aqi_level": document["no2"]["aqi_level"],
+                "pm10_aqi_level": document["pm10"]["aqi_level"],
+                "pm2_5_aqi_level": document["pm2_5"]["aqi_level"],
+            }
+            database_dictionary_list.append(document_as_dict)
+
+    except ConnectionFailure:
+        raise RuntimeError("Failed to connect to the database")
+    except OperationFailure as e:
+        raise RuntimeError(f"Database operation failed: {e}")
+
+    finally:
+        client.close()
+
     return database_dictionary_list
 
 
@@ -153,7 +174,7 @@ def export_data_array_to_excel(data_array: xarray.DataArray, filename: str):
 
 
 def export_cams_data_to_excel_by_level(
-    single_level_dataset: xarray.Dataset, multi_level_dataset: xarray.Dataset
+        single_level_dataset: xarray.Dataset, multi_level_dataset: xarray.Dataset
 ):
     print("\nExporting to Excel: Single level data...")
     export_dataset_to_excel(single_level_dataset, "AllSingleLevelData.xlsx")
@@ -162,7 +183,7 @@ def export_cams_data_to_excel_by_level(
 
 
 def export_to_excel_by_pollutant(
-    single_level_dataset: xarray.Dataset, multi_level_dataset: xarray.Dataset
+        single_level_dataset: xarray.Dataset, multi_level_dataset: xarray.Dataset
 ):
     print("\nExporting to Excel...")
     export_data_array_to_excel(single_level_dataset["pm10"], "pm10.xlsx")
@@ -173,7 +194,7 @@ def export_to_excel_by_pollutant(
 
 
 def longitude_calculator_for_cams_data(
-    longitude_value: float, data_increment_size: float
+        longitude_value: float, data_increment_size: float
 ):
     if -180 < longitude_value <= 0 - (data_increment_size / 2):
         return float(Decimal(str(longitude_value)) + Decimal("360"))
@@ -195,8 +216,8 @@ def convert_cams_locations_file_to_dict(cams_locations_file_name: str) -> list[d
 
 
 def get_ecmwf_forecast_to_dict_for_countries(
-    cams_locations_file_name: str,
-    ecmwf_forecast_file_name: str,
+        cams_locations_file_name: str,
+        ecmwf_forecast_file_name: str,
 ):
     ecmwf_countries_dict = convert_cams_locations_file_to_dict(cams_locations_file_name)
     list_of_records = read_csv(ecmwf_forecast_file_name).to_dict("records")
@@ -210,11 +231,11 @@ def get_ecmwf_forecast_to_dict_for_countries(
 
 
 def calculate_database_divergence_from_ecmwf_forecast_values(
-    database_ozone_value: float, ecmwf_forecast_ozone_value: float
+        database_ozone_value: float, ecmwf_forecast_ozone_value: float
 ) -> float:
     divergence_percentage = (
-        (database_ozone_value - ecmwf_forecast_ozone_value) / ecmwf_forecast_ozone_value
-    ) * 100
+                                    (database_ozone_value - ecmwf_forecast_ozone_value) / ecmwf_forecast_ozone_value
+                            ) * 100
     if divergence_percentage < 0:
         formatted_divergence_percentage = divergence_percentage * -1
     else:
@@ -223,35 +244,35 @@ def calculate_database_divergence_from_ecmwf_forecast_values(
 
 
 def get_ecmwf_record_for_city_and_valid_time(
-    test_city: str, test_forecast_valid_time: datetime, ecmwf_all_data: list[dict]
+        test_city: str, test_forecast_valid_time: datetime, ecmwf_all_data: list[dict]
 ) -> list[dict]:
     return list(
         filter(
             lambda x: x["location_name"] == test_city
-            and x["valid_time"] == test_forecast_valid_time.strftime("%Y-%m-%dT%H:%M"),
+                      and x["valid_time"] == test_forecast_valid_time.strftime("%Y-%m-%dT%H:%M"),
             ecmwf_all_data,
         )
     )
 
 
 def get_database_record_for_city_and_valid_time(
-    test_forecast_base_time: datetime,
-    test_city: str,
-    test_forecast_valid_time: datetime,
-    database_all_data: list[dict[str]],
+        test_forecast_base_time: datetime,
+        test_city: str,
+        test_forecast_valid_time: datetime,
+        database_all_data: list[dict[str]],
 ) -> list[dict]:
     return list(
         filter(
             lambda x: x["forecast_base_time"] == test_forecast_base_time
-            and x["name"] == test_city
-            and x["forecast_valid_time"] == test_forecast_valid_time,
+                      and x["name"] == test_city
+                      and x["forecast_valid_time"] == test_forecast_valid_time,
             database_all_data,
         )
     )
 
 
 def get_pollutant_value(
-    pollutant: str, source_name: str, list_of_records_from_source: list[dict]
+        pollutant: str, source_name: str, list_of_records_from_source: list[dict]
 ):
     first_record = list_of_records_from_source[0]
     pollutant_name_upper_case = pollutant.upper()
@@ -273,12 +294,12 @@ def get_pollutant_value(
 
 
 def get_forecast_percentage_divergence(
-    test_city: str,
-    test_forecast_valid_time: datetime,
-    ecmwf_all_data: list[dict],
-    test_forecast_base_time: datetime,
-    database_all_data: list[dict[str]],
-    pollutant: str,
+        test_city: str,
+        test_forecast_valid_time: datetime,
+        ecmwf_all_data: list[dict],
+        test_forecast_base_time: datetime,
+        database_all_data: list[dict[str]],
+        pollutant: str,
 ) -> float:
     ecmwf_record_for_city_and_valid_time = get_ecmwf_record_for_city_and_valid_time(
         test_city, test_forecast_valid_time, ecmwf_all_data
